# docs/openapi/reconcile-v1.yaml
openapi: 3.1.0
info:
  title: Reconcile - payment & settlement reconciliation
  description: |
    Money movement, provider settlement ingestion, and reconciliation.

    Every mutating endpoint requires an `Idempotency-Key` and replays a
    stored response for 24 hours. Refusals are RFC 9457 problem
    documents; the failure taxonomy is in `docs/spec/07-failure-matrix.md`
    and the error table in `docs/spec/06-api-contract.md`.

    Two routes are deliberately unauthenticated: `GET /api/v1/health`,
    which a load balancer must be able to reach, and
    `POST /api/v1/provider/webhooks`, which authenticates by HMAC
    signature rather than by token.
  license:
    name: Proprietary
  version: v1
servers:
- url: /
  description: This application
security:
- bearerAuth: []
paths:
...
  /api/v1/audit/events:
    get:
      tags:
      - operations-controller
      operationId: auditEvents
      parameters:
      - name: action
        in: query
        required: false
        schema:
          type: string
      - name: entityType
        in: query
        required: false
        schema:
          type: string
      - name: entityId
        in: query
        required: false
        schema:
          type: string
      - name: actorType
        in: query
        required: false
        schema:
          type: string
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: requestId
        in: query
        required: false
        schema:
          type: string
      - name: after
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: afterId
        in: query
        required: false
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEventView"
  /api/v1/audit/entities/{type}/{id}:
    get:
      tags:
      - operations-controller
      summary: "Audit trail for a single entity, by entity type and id"
      operationId: entityTrail
      parameters:
      - name: type
        in: path
        description: "Entity type token. Case-insensitive: `payment` and `PAYMENT`\
          \ are the same request. Any other value is a 404. Accepted: PAYMENT, PAYOUT_RECORD,\
          \ PROVIDER, PROVIDER_EVENT, SETTLEMENT_RECORD, RECONCILIATION_BATCH, RECONCILIATION_RESULT,\
          \ RECONCILIATION_CASE, IDEMPOTENCY_RECORD."
        required: true
        schema:
          type: string
      - name: id
        in: path
        description: "The entity's id, as stored in `audit_events.entity_id`"
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEventView"
...
  /api/v1/accounts/{code}:
    get:
      tags:
      - ledger-controller
      operationId: account
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      - name: currency
        in: query
        required: false
        schema:
          type: string
          default: EGP
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AccountBalance"
...
components:
  schemas:
...
    MoneyDto:
      type: object
      properties:
        amountMinor:
          type: integer
          format: int64
        currency:
          type: string
...
    AuditEventView:
      type: object
      properties:
        id:
          type: string
        occurredAt:
          type: string
          format: date-time
        actorType:
          type: string
        actorId:
          type: string
        action:
          type: string
        entityType:
          type: string
        entityId:
          type: string
        requestId:
          type: string
        metadata:
          type: string
...
    AccountBalance:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        accountType:
          type: string
        currency:
          type: string
          enum:
          - EGP
          - USD
          - EUR
          - GBP
        balance:
          $ref: "#/components/schemas/MoneyDto"
        recomputedBalance:
          $ref: "#/components/schemas/MoneyDto"
        entryCount:
          type: integer
          format: int64
        recomputedEntryCount:
          type: integer
          format: int64
        lastEntryAt:
          type: string
          format: date-time
        verified:
          type: boolean
...
  securitySchemes:
    bearerAuth:
      type: http
      description: |
        A static operator or admin bearer token. Admin implies
        operator. Not a production identity design; see the
        authentication section of the README for what a real
        deployment puts in front of this instead.
      scheme: bearer